LayoutTests/imported/w3c/web-platform-tests/dom/ranges/Range-extractContents-dynamic-end-expected.txt

Middle elements
Should NOT be extracted

PASS Range.extractContents should not extract out-of-bounds nodes if end container is removed during iframe unload
